Fix foxglove.Grid default auto-select colorMode behavior (#5350)
**User-Facing Changes**
<!-- will be used as a changelog entry -->
- Fix foxglove.Grid default auto-select colorMode behavior to use
separate RGBA-fields when available

**Description**
`autoSelectColorField` formerly did not account for separate RGBA
fields. Now it should.

<!-- link relevant GitHub issues -->
Fixes FG-1718

<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->
